+++ gems/aws-sdk-s3/spec/multipart_file_uploader_spec.rb
@@ -126,20 +126,11 @@ module Aws
         end
 
         it 'reports when it is unable to abort a failed multipart upload' do
-          client.stub_responses(
-            :upload_part,
-            [
-              { etag: 'etag-1' },
-              { etag: 'etag-2' },
-              { etag: 'etag-3' },
-              RuntimeError.new('part failed')
-            ]
-          )
-          client.stub_responses(:abort_multipart_upload, [RuntimeError.new('network-error')])
+          client.stub_responses(:upload_part, RuntimeError.new('part failed'))
+          client.stub_responses(:abort_multipart_upload, RuntimeError.new('network-error'))
           expect do
             subject.upload(large_file, params)
           end.to raise_error(/failed to abort multipart upload: network-error. Multipart upload failed: part failed/)
-
         end
 
         it 'aborts multipart upload when upload fails to complete' do