+++ gems/aws-sdk-s3/spec/plugins/checksum_algorithm_spec.rb
@@ -3,8 +3,8 @@
 require_relative '../spec_helper'
 
 module Aws
-  module S3
-    describe Client do
+  module Plugins
+    describe ChecksumAlgorithm do
       let(:creds) { Aws::Credentials.new('akid', 'secret') }
       let(:client) { S3::Client.new(stub_responses: true) }
       let(:bucket) { 'bucket' }
@@ -20,10 +20,10 @@ module Aws
         client.stub_responses(
           :get_object,
           [{
-             body: body,
-             headers: {'x-amz-checksum-crc32' => digest},
-             status_code: 200
-           }]
+            body: body,
+            headers: { 'x-amz-checksum-crc32' => digest },
+            status_code: 200
+          }]
         )
         resp = client.get_object(bucket: bucket, key: key)
         expect(resp.context[:http_checksum][:validated]).to eq 'CRC32'
@@ -33,10 +33,10 @@ module Aws
         client.stub_responses(
           :get_object,
           [{
-             body: body,
-             headers: {'x-amz-checksum-crc32' => 'invalid_value'},
-             status_code: 200
-           }]
+            body: body,
+            headers: { 'x-amz-checksum-crc32' => 'invalid_value' },
+            status_code: 200
+          }]
         )
         expect do
           client.get_object(bucket: bucket, key: key)
@@ -47,12 +47,12 @@ module Aws
         client.stub_responses(
           :get_object,
           [{
-             body: body_part_1,
-             headers: {'x-amz-checksum-crc32' => digest_part_1},
-             status_code: 200
-           }]
+            body: body_part_1,
+            headers: { 'x-amz-checksum-crc32' => digest_part_1 },
+            status_code: 200
+          }]
         )
-        resp = client.get_object(bucket: bucket, key: key, range: "bytes=0-6")
+        resp = client.get_object(bucket: bucket, key: key, range: 'bytes=0-6')
         expect(resp.context[:http_checksum][:validated]).to eq 'CRC32'
       end
 
@@ -60,10 +60,10 @@ module Aws
         client.stub_responses(
           :get_object,
           [{
-             body: body_part_1,
-             headers: {'x-amz-checksum-crc32' => digest_part_1},
-             status_code: 200
-           }]
+            body: body_part_1,
+            headers: { 'x-amz-checksum-crc32' => digest_part_1 },
+            status_code: 200
+          }]
         )
         resp = client.get_object(bucket: bucket, key: key, part_number: 1)
         expect(resp.context[:http_checksum][:validated]).to eq 'CRC32'
@@ -100,13 +100,85 @@ module Aws
             client.stub_responses(
               :get_object,
               [{
-                 body: test_case['responsePayload'],
-                 headers: test_case['responseHeaders'],
-                 status_code: 200
+                body: test_case['responsePayload'],
+                headers: test_case['responseHeaders'],
+                status_code: 200
               }]
             )
             resp = client.get_object(bucket: bucket, key: key, checksum_mode: 'ENABLED')
             expect(resp.context[:http_checksum][:validated]).to be_nil
+          end
+        end
+      end
+
+      context 'AwsChunkedTrailerDigestIO' do
+        let(:subject) do
+          ChecksumAlgorithm::AwsChunkedTrailerDigestIO.new(
+            io: StringIO.new('x' * (32 * 1024)),
+            algorithm: 'CRC32',
+            location_name: 'x-amz-checksum-crc32'
+          )
+        end
+
+        describe '#size' do
+          it 'matches actual read output length' do
+            expect(subject.size).to eq(subject.read.bytesize)
+          end
+
+          it 'returns zero bytes for empty content' do
+            expected_empty_trailer = "0\r\nx-amz-checksum-crc32:AAAAAA==\r\n\r\n".bytesize
+            subject = ChecksumAlgorithm::AwsChunkedTrailerDigestIO.new(
+              io: StringIO.new(''),
+              algorithm: 'CRC32',
+              location_name: 'x-amz-checksum-crc32'
+            )
+            expect(subject.size).to eq(expected_empty_trailer)
+          end
+        end
+
+        describe '#read' do
+          it 'returns full chunked output when no length given' do
+            output = subject.read
+            expect(output.scan("3ff8\r\n").length).to eq(2)
+            expect(output).to end_with("0\r\nx-amz-checksum-crc32:uj4bwQ==\r\n\r\n")
+          end
+
+          it 'returns partial data when length specified' do
+            partial = subject.read(10)
+            expect(partial.bytesize).to eq(10)
+            expect(partial).to start_with("3ff8\r\n")
+          end
+
+          it 'maintains content integrity across mixed read patterns' do
+            chunks = []
+            # Read in random chunks until EOF
+            while (chunk = subject.read(rand(100..2000)))
+              chunks << chunk
+            end
+
+            subject.rewind
+            expect(chunks.join).to eq(subject.read)
+          end
+        end
+
+        describe '#rewind' do
+          it 'resets position to beginning' do
+            first_read = subject.read(100)
+            subject.rewind
+            second_read = subject.read(100)
+
+            expect(first_read).to eq(second_read)
+          end
+        end
+
+        describe '#eof?' do
+          it 'returns false at start of IO' do
+            expect(subject.eof?).to be false
+          end
+
+          it 'returns true after reading all data' do
+            subject.read
+            expect(subject.eof?).to be true
           end
         end
       end