+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,7 +1,7 @@
 Unreleased Changes
 ------------------
 
-* Issue - Fix flaky login credentials spec test.
+* Feature - Improved memory efficiency when calculating request checksums.
 
 3.240.0 (2025-12-16)
 ------------------